requirements.txt
amqp==2.5.0
billiard==3.6.0.0
celery==4.3.0
configparser==3.7.4
Django==2.2.3
django-haystack==2.8.1
django-redis==4.10.0
django-tinymce==2.8.0
itsdangerous==1.1.0
jieba==0.39
kombu==4.6.3
mysqlclient==1.4.2.post1
Pillow==6.1.0
py3Fdfs==2.2.0
pycryptodomex==3.7.2
python-alipay-sdk==1.10.1
pytz==2019.1
redis==3.2.1
sqlparse==0.3.0
uWSGI==2.0.18
vine==1.3.0
Whoosh==2.7.4
